+++ eggs/naaya.content.talkback/CHANGELOG.rst
@@ -1,5 +1,6 @@
 1.5.1 (2019-01-24)
 ------------------
+* fix for the comments count for invited commenters [dumitval]
 * fixing update script [batradav]
 
 1.5.0 (2019-01-24)